+++ lib.esm/providers/index.js
@@ -13,7 +13,7 @@ export { AbstractProvider, UnmanagedSubscriber } from "./abstract-provider.js";
 export { AbstractSigner, VoidSigner, } from "./abstract-signer.js";
 export { showThrottleMessage } from "./community.js";
 export { getDefaultProvider } from "./default-provider.js";
-export { EnsResolver } from "./ens-resolver.js";
+export { EnsResolver, MulticoinProviderPlugin } from "./ens-resolver.js";
 export { Network } from "./network.js";
 export { NonceManager } from "./signer-noncemanager.js";
 export { NetworkPlugin, GasCostPlugin, EnsPlugin, FeeDataNetworkPlugin, } from "./plugins-network.js";