+++ lib.esm/providers/provider-jsonrpc.js
@@ -60,7 +60,8 @@ const defaultOptions = {
     staticNetwork: null,
     batchStallTime: 10,
     batchMaxSize: (1 << 20),
-    batchMaxCount: 100 // 100 requests
+    batchMaxCount: 100,
+    cacheTimeout: 250
 };
 // @TODO: Unchecked Signers
 export class JsonRpcSigner extends AbstractSigner {
@@ -278,7 +279,11 @@ export class JsonRpcApiProvider extends AbstractProvider {
         }, stallTime);
     }
     constructor(network, options) {
-        super(network);
+        const superOptions = {};
+        if (options && options.cacheTimeout != null) {
+            superOptions.cacheTimeout = options.cacheTimeout;
+        }
+        super(network, superOptions);
         this.#nextId = 1;
         this.#options = Object.assign({}, defaultOptions, options || {});
         this.#payloads = [];